# rose-backend-main/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/cart/add": {
      "post": {
        "operationId": "CartController_addItem",
        "parameters": [
          {
            "name": "x-session-id",
            "in": "header",
            "description": "Session ID for the cart",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/AddToCartDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddToCartDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Cart"
        ]
      }
    },
    "/cart/add-multiple": {
      "post": {
        "operationId": "CartController_addMultiple",
        "parameters": [
          {
            "name": "x-session-id",
            "in": "header",
            "description": "Session ID for the cart",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/x-www-form-urlencoded": {
              "schema": {
                "$ref": "#/components/schemas/AddMultipleItemsDto"
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddMultipleItemsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Cart"
        ]
      }
    },
    "/cart": {
      "get": {
        "operationId": "CartController_getCart",
        "parameters": [
          {
            "name": "x-session-id",
            "in": "header",
            "description": "Session ID for the cart",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Cart"
        ]
      }
    },
    "/cart/remove/{id}": {
      "delete": {
        "operationId": "CartController_removeItem",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "x-session-id",
            "in": "header",
            "description": "Session ID for the cart",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Cart"
        ]
      }
    },
    "/cart/clear": {
      "delete": {
        "operationId": "CartController_clearCart",
        "parameters": [
          {
            "name": "x-session-id",
            "in": "header",
            "description": "Session ID for the cart",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Cart"
        ]
      }
    },
    "/admin/orders": {
      "get": {
        "operationId": "OrderAdminController_findAll",
        "parameters": [
          {
            "name": "phone",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "OrderAdmin"
        ]
      }
    },
    "/admin/orders/{id}": {
      "get": {
        "operationId": "OrderAdminController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "OrderAdmin"
        ]
      },
      "delete": {
        "operationId": "OrderAdminController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "Authorization": []
          }
        ],
        "tags": [
          "OrderAdmin"
        ]
      }
    },
...
  },
  "info": {
    "title": "YouTaab API",
    "description": "API for coffeshop reservation and menu",
    "version": "0.0.1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "Authorization": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "AddToCartDto": {
        "type": "object",
        "properties": {
          "menuItemId": {
            "type": "number"
          },
          "quantity": {
            "type": "number"
          }
        },
        "required": [
          "menuItemId",
          "quantity"
        ]
      },
      "AddMultipleItemsDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AddToCartDto"
            }
          }
        },
        "required": [
          "items"
        ]
      },
...
    }
  }
}